mysite/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
cssselect==1.2.0
cssutils==2.11.1
daphne==4.1.2
Django==4.2.14
djangorestframework==3.15.2
emails==0.6
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
lxml==5.2.2
more-itertools==10.3.0
msgpack==1.0.8
mysqlclient==2.2.4
premailer==3.10.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyMySQL==1.1.1
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
redis==5.0.8
requests==2.32.3
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.2
zope.interface==6.4.post2